+++ server/controllers/crm/crm4Controller.ts
@@ -1,10 +1,10 @@
 import type { Request, RequestHandler, Response } from 'express'
 import { Crm4Response } from '@crm4'
-import CrmApiService from '../services/crmApiService'
-import NavigationService from '../services/navigationService'
-import CrmDisplayService from '../services/crmDisplayService'
+import CrmApiService from '../../services/crmApiService'
+import NavigationService from '../../services/navigationService'
+import CrmDisplayService from '../../services/crmDisplayService'
 
-export default class Crm5Controller {
+export default class Crm4Controller {
   constructor(
     private readonly crm4Service: CrmApiService<Crm4Response>,
     private readonly navigationService: NavigationService,
@@ -13,6 +13,8 @@ export default class Crm5Controller {
 
   show(): RequestHandler {
     return async (req: Request, res: Response): Promise<void> => {
+      const usn = Number(req.params.usn)
+      const crm4Response = await this.crm4Service.getCrm(usn)
       res.render('pages/crmDetails', { title: 'CRM4', navigationItems: {}, section: {} })
     }
   }
